% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{addAnnotationsToDataSet}
\alias{addAnnotationsToDataSet}
\title{Add Annotations To DESeqDataSet}
\usage{
addAnnotationsToDataSet(dds, ann)
}
\arguments{
\item{dds}{A \link[DESeq2]{DESeqDataSet} object.}

\item{ann}{A \link{data.frame} containing gene symbols and/or delimited GO
terms.}
}
\value{
A \link[DESeq2]{DESeqDataSet} object with annotations.
}
\description{
\code{addAnnotationsToDataSet} adds annotations from a dataframe
  containing gene symbols and/or GO terms to the \link[DESeq2]{DESeqDataSet}
  object.
}
\details{
The annotation dataframe \code{ann} contains gene symbols and/or GO
  terms associated with the genes in the \link[DESeq2]{DESeqDataSet} object.
  Each row in the dataframe represents a gene. The first column contains the
  identifiers which correspond to the rownames of the
  \link[DESeq2]{DESeqDataSet} object.

  The first column must contain the gene identifiers which correspond to the
  row names of the \link[DESeq2]{DESeqDataSet} object. There doesn't need to
  be an entry for every gene identifier in the \link[DESeq2]{DESeqDataSet}
  but, at least one identifier should be represented.

  There then follows one or two further columns depending on whether the
  annotations consist of only gene symbols, only GO terms or both gene
  symbols and GO terms. Where there are multiple GO terms for a gene, the
  individual terms should be semi-colon (;) delimited e.g.
  \emph{GO:0004415;GO:0004415;GO:0004415;}.

  \code{addAnnotationsToDataSet} returns a \link[DESeq2]{DESeqDataSet} with
  annotations stored in the object metadata. Gene symbols will be accessible
  using \code{mcols(dds)$symbol} while GO terms will be accessible using
  \code{metadata(dds)$go} where dds represents a \link[DESeq2]{DESeqDataSet}
  object.
}
\examples{
library(DESeq2)
dds <- makeExampleDESeqDataSet(n=3)
good_ann <- data.frame( genes=rownames(dds),
                        gene_names=letters[1:3],
                        go=c("GO:0000001", "GO:0000001;GO:0000002", ""))
good_ann[] <- lapply(good_ann, as.character)
addAnnotationsToDataSet(dds, good_ann)
}
\seealso{
Other annotation functions: \code{\link{annotateDataset}},
  \code{\link{getGOlist}}, \code{\link{getGeneSymbols}},
  \code{\link{importAnnotation}}
}
