% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotateDataset}
\alias{annotateDataset}
\title{Annotate DESeqDataSet}
\usage{
annotateDataset(dds, parameters)
}
\arguments{
\item{dds}{A \link[DESeq2]{DESeqDataSet} object.}

\item{parameters}{A \link{list} containing key/value pairs which define the
parameters for the analysis (see \link[deago]{importConfig} for more
information).}
}
\value{
A \link[DESeq2]{DESeqDataSet} object with annotations.
}
\description{
\code{annotateDataset} reads a tab-delimited annotation file
  containing gene symbols and/or GO terms into a dataframe which is used to
  annotate the DESeqDataSet.
}
\details{
The annotation file is a text file containing gene information. Rows
  correspond to genes while columns correspond to gene symbols and GO terms
  that are associated with the genes. The annotation file should \emph{not}
  contain a header row.

  The first column must contain the gene identifiers which correspond to the
  row names of the \link[DESeq2]{DESeqDataSet} object. There doesn't need to
  be an entry for every gene identifier in the \link[DESeq2]{DESeqDataSet}
  but, at least one identifier should be represented.

  There then follows one or two further columns depending on whether the
  annotations consist of only gene symbols, only GO terms or both gene
  symbols and GO terms. Where there are multiple GO terms for a gene, the
  individual terms should be semi-colon (;) delimited e.g.
  \emph{GO:0004415;GO:0004415;GO:0004415;}.

  \code{annotateDataset} returns a \link[DESeq2]{DESeqDataSet} with the
  annotations stored in the object metadata. Gene symbols will be accessible
  using \code{mcols(dds)$symbol} while GO terms will be accessible using
  \code{metadata(dds)$go} where \code{dds} represents a
  \link[DESeq2]{DESeqDataSet} object.
}
\examples{
\dontrun{
annotateDataset(dds, parameters)
}
}
\seealso{
Other annotation functions: \code{\link{addAnnotationsToDataSet}},
  \code{\link{getGOlist}}, \code{\link{getGeneSymbols}},
  \code{\link{importAnnotation}}

Other import functions: \code{\link{importAnnotation}},
  \code{\link{importConfig}}, \code{\link{importTargets}},
  \code{\link{readCountData}},
  \code{\link{validateConfig}},
  \code{\link{validateTargets}}
}
\concept{annotation functions}
\concept{import functions}
