% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{buildConfig}
\alias{buildConfig}
\title{Build Configuration File}
\usage{
buildConfig(file, path = NULL, parameters)
}
\arguments{
\item{file}{A \link{character} string giving the name of the file.}

\item{path}{A \link{character} string giving the directory containing the
file. This may be omitted if the file is in the current working directory.}

\item{parameters}{A \link{list} containing key/value pairs which define the
parameters for the analysis.}
}
\value{
A \link{character} string containing the name of the output configuration file.
}
\description{
\code{buildConfig} writes a list of validated key/value pairs to
  a tab-delimited file.
}
\details{
\code{buildConfig} takes a list of key/value pairs which represent
  the parameters for the analysis, validates them using
  \code{\link{validateConfig}} and then writes them to a tab-delimited
  configuration file using \code{\link{writeConfig}}.

  See \code{\link{importConfig}} for required and accepted key/value pairs.

  \code{buildConfig} returns the new configuration file location.
}
\examples{
\dontrun{
 # Basic parameter list
 params <- list ( 'counts_directory' = "/path/to/counts.txt",
                  'targets_file'     = "/path/to/targets.txt",
                  'results_directory' = "/path/to/result_dir")

 # Write tab-delimited config file to working directory
 buildConfig("config.txt", parameters=params)

 # Write tab-delimited config file to another directory
 buildConfig("config.txt", path="/path/to/config", parameters=params)
}
}
\seealso{
Other configuration functions: \code{\link{importConfig}},
  \code{\link{validateConfig}}, \code{\link{writeConfig}}
}
\concept{configuration functions}
