% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseqAnalysis.R
\name{constructDESeqDataset}
\alias{constructDESeqDataset}
\title{Construct DESeq2 dataset}
\usage{
constructDESeqDataset(countdata, coldata)
}
\arguments{
\item{countdata}{Name of object containing count data}

\item{coldata}{input matrix for DESeq2 analysis}
}
\value{
Returns a DESeqDataSet object
}
\description{
A DESeqDataSet is a subclass of RangedSummarizedExperiment, used to store the input values, intermediate calculations and results of an analysis of differential expression.
}
