% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasts.R
\name{contrastSummary}
\alias{contrastSummary}
\title{Differential expression summary}
\usage{
contrastSummary(contrastList, parameters)
}
\arguments{
\item{contrastList}{list of contrast results objects}

\item{parameters}{list of parameters}
}
\description{
Summary table of differentially expressed gene number per contrast
}
