% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goAnalysis.R
\name{filterGeneIdentifiers}
\alias{filterGeneIdentifiers}
\title{Filter DE gene identifiers in GO term}
\usage{
filterGeneIdentifiers(contrast, goGeneList, gene_level = "all")
}
\arguments{
\item{contrast}{DESeqResults object}

\item{goGeneList}{vector of gene identifiers associated with GO term}

\item{gene_level}{all, up or down [all]}
}
\description{
Filter DE gene identifiers for GO terms
}
