% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goAnalysis.R
\name{filterGeneSymbols}
\alias{filterGeneSymbols}
\title{Filter DE gene symbols in GO term}
\usage{
filterGeneSymbols(contrast, goGeneList, gene_level = "all")
}
\arguments{
\item{contrast}{DESeqResults object}

\item{goGeneList}{vector of symbols associated with GO term}

\item{gene_level}{all, up or down [all]}
}
\description{
Filter DE gene symbols for GO terms
}
