% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcPlots.R
\name{formatReadCounts}
\alias{formatReadCounts}
\title{Format count matrix from DESeqDataSet}
\usage{
formatReadCounts(dds)
}
\arguments{
\item{dds}{DESeqDataSet}
}
\description{
Format counts from DESeqDataSet for read and null counts plots
}
