% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasts.R
\name{getContrasts}
\alias{getContrasts}
\title{Extract contrasts from DESeq2 results}
\usage{
getContrasts(dds, parameters)
}
\arguments{
\item{dds}{A \link[DESeq2]{DESeqDataSet} object.}

\item{parameters}{A \link{list} containing key/value pairs which define the
parameters for the analysis (see \link[deago]{importConfig} for more
information).}
}
\description{
Extract contrasts from DESeq2 results and write to file
}
\details{
The details
}
