% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{getGeneSymbols}
\alias{getGeneSymbols}
\title{Extract Gene Symbols}
\usage{
getGeneSymbols(dds, ann, col)
}
\arguments{
\item{dds}{A \link[DESeq2]{DESeqDataSet} object.}

\item{ann}{A \link{data.frame} containing gene symbols and/or delimited GO
terms.}

\item{col}{An \link{integer} representing the index of the column containing
gene symbols.}
}
\value{
A vector of gene symbols corresponding to the rownames of the \link[DESeq2]{DESeqDataSet} object.
}
\description{
\code{getGeneSymbols} extracts a vector of gene symbols from a
  dataframe.
}
\details{
\code{getGeneSymbols} requires an annotation dataframe \code{ann}
  and an integer value \code{col} which represents the index of the column
  containing the gene symbols.  For example, if the gene symbols were in
  column 2: \code{getGeneSymbols(ann,2)}.

  At least one of the gene identifiers in the row names of the
  \link[DESeq2]{DESeqDataSet} object must be present in the first column of
  the annotation dataframe \code{ann}.

  Gene symbols are extracted from the dataframe and any missing identifiers
  are given a gene symbol of "unknown". There is a final check to ensure that
  the gene symbol vector is the same length as the row names vector from the
  \link[DESeq2]{DESeqDataSet} object.

  \code{getGeneSymbols} returns a vector of gene symbols associated with the
  gene identifiers of the \link[DESeq2]{DESeqDataSet} object.
}
\examples{
library(DESeq2)
dds <- makeExampleDESeqDataSet(n=3)
good_ann <- data.frame( genes=rownames(dds),
                        gene_names=letters[1:3],
                        go=c("GO:0000001", "GO:0000001;GO:0000002", ""))
good_ann[] <- lapply(good_ann, as.character)
getGeneSymbols(dds, good_ann, 2)
}
\seealso{
Other annotation functions: \code{\link{addAnnotationsToDataSet}},
  \code{\link{annotateDataset}}, \code{\link{getGOlist}},
  \code{\link{importAnnotation}}
}
\concept{annotation functions}
