% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goAnalysis.R
\name{getGOidentifiers}
\alias{getGOidentifiers}
\title{Get gene identifiers for GO terms}
\usage{
getGOidentifiers(GOdata, GOtable)
}
\arguments{
\item{GOdata}{topGO object}

\item{GOtable}{topGO table}
}
\description{
Get gene identifiers for GO terms
}
