% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{getGOlist}
\alias{getGOlist}
\title{Transform Delimited GO Terms To List}
\usage{
getGOlist(ann, col)
}
\arguments{
\item{ann}{A \link{data.frame} containing gene symbols and/or delimited GO
terms.}

\item{col}{An \link{integer} representing the index of the column containing
GO terms.}
}
\value{
A \link{list} of GO terms where names are the associated gene
  identifiers.
}
\description{
\code{getGOlist} transforms a semi-colon delimited \link{vector}
  of GO terms and their corresponding gene identifiers from a
  \link{data.frame} into a \link{list}.
}
\details{
The annotation dataframe \code{ann} contains gene symbols and/or GO
  terms associated with the genes in the \link[DESeq2]{DESeqDataSet} object.
  Each row in the dataframe represents a gene. The first column contains the
  identifiers which correspond to the rownames of the
  \link[DESeq2]{DESeqDataSet} object.

  There then follows one or two further columns depending on whether the
  annotations consist of only gene symbols, only GO terms or both gene
  symbols and GO terms.  For more information see
  \link[deago]{annotateDataset}.

  \code{getGOlist} requires an annotation dataframe \code{ann} and an integer
  value \code{col} which represents the index of the column containing the GO
  terms.  For example, if the GO terms were in column 2:
  \code{getGOlist(ann,2)}.

  Where there are multiple GO terms for a gene, the individual terms should
  be semi-colon (;) delimited e.g. \emph{GO:0004415;GO:0004415;GO:0004415;}.

  \code{getGOlist} returns a list where the names are the gene idenifiers and
  the values are the associated GO terms.  This format is required for
  \link{topGO} GO term enrichment analysis.
}
\examples{
library(DESeq2)
dds <- makeExampleDESeqDataSet(n=3)
good_ann <- data.frame( genes=rownames(dds),
                        gene_names=letters[1:3],
                        go=c("GO:0000001", "GO:0000001;GO:0000002", ""))
good_ann[] <- lapply(good_ann, as.character)
getGOlist(good_ann, 3)
}
\seealso{
Other annotation functions: \code{\link{addAnnotationsToDataSet}},
  \code{\link{annotateDataset}},
  \code{\link{getGeneSymbols}},
  \code{\link{importAnnotation}}
}
\concept{annotation functions}
