% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcPlots.R
\name{getPrincipalComponents}
\alias{getPrincipalComponents}
\title{Calculate principal components}
\usage{
getPrincipalComponents(dds)
}
\arguments{
\item{dds}{DESeqDataSet}
}
\description{
Calcultate principal component values
}
