% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrastPlots.R
\name{getTopGenes}
\alias{getTopGenes}
\title{Subset top DE genes}
\usage{
getTopGenes(df, n)
}
\arguments{
\item{df}{Dataframe from a \link{DESeqResults} object}

\item{n}{Number of genes to label in each direction}
}
\description{
Helper function to extract n DE genes with the greatest
  Log2FoldChange and n genes with the lowest Log2FoldChange from a
  DESeqResults dataframe.
}
