% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrastPlots.R
\name{getVennCounts}
\alias{getVennCounts}
\title{Generate count matrix for venn diagram}
\usage{
getVennCounts(contrastList)
}
\arguments{
\item{contrastList}{List of contrasts (pairwise comparisons)}
}
\description{
Generate count matrix for venn diagram for DE genes
}
