% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{importAnnotation}
\alias{importAnnotation}
\title{Import Annotation File}
\usage{
importAnnotation(file)
}
\arguments{
\item{file}{A \link{character} string giving the name of the annotation file.}
}
\value{
A dataframe with annotations.
}
\description{
\code{importAnnotation} reads a tab-delimited annotation file
  containing gene symbols and/or GO terms into a dataframe.
}
\details{
The annotation file is a text file containing gene information. Rows
  correspond to genes while columns correspond to gene symbols and GO terms
  that are associated with the genes. The annotation file should \emph{not}
  contain a header row.

  The first column must contain the gene identifiers which correspond to the
  row names of the \link[DESeq2]{DESeqDataSet} object. There doesn't need to
  be an entry for every gene identifier in the \link[DESeq2]{DESeqDataSet}
  but, at least one identifier should be represented.

  There then follows one or two further columns depending on whether the
  annotations consist of only gene symbols, only GO terms or both gene
  symbols and GO terms. Where there are multiple GO terms for a gene, the
  individual terms should be semi-colon (;) delimited e.g.
  \emph{GO:0004415;GO:0004415;GO:0004415;}.
}
\examples{
\dontrun{
importAnnotaton(annotation_file)
}
}
\seealso{
Other annotation functions: \code{\link{addAnnotationsToDataSet}},
  \code{\link{annotateDataset}}, \code{\link{getGOlist}},
  \code{\link{getGeneSymbols}}

Other import functions: \code{\link{annotateDataset}},
  \code{\link{importConfig}}, \code{\link{importTargets}},
  \code{\link{readCountData}},
  \code{\link{validateConfig}},
  \code{\link{validateTargets}}
}
