% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{importConfig}
\alias{importConfig}
\title{Read Configuration File}
\usage{
importConfig(file, path = NULL, sep = "\\t")
}
\arguments{
\item{file}{A \link{character} string giving the name of the file.}

\item{path}{A \link{character} string giving the directory containing the
file. This may be omitted if the file is in the current working directory.}

\item{sep}{A \link{character} used to separate the fields.}
}
\value{
A key/value \link{list} containing the parameters from the configuration file.
}
\description{
\code{importConfig} reads configuration file into a list.
}
\details{
\code{importConfig} reads a tab-delimited configuration file
  containing key/value pairs which represent the parameters for the analysis
  into a list.  The configuration file should contain two columns, the first
  containing the keys outlined below and the second, their values.

  Required key/value pairs are: \describe{ \item{\strong{counts_directory}}{Directory
  where count data files are located.} \item{\strong{targets_file}}{Location of
  targets file containing the mappings between file names and experimental
  conditions.} \item{\strong{results_directory}}{Location where results directory
  is to be created.} }

  Other accepted key/value pairs and their defaults are: \describe{
  \item{\strong{gene_ids}}{Name of column in count files which contains the
  gene identifiers [Default: Geneid]} \item{\strong{alpha}}{Significance
  cut-off, see \link[DESeq2]{results} for more information [Default: 0.05]}
  \item{\strong{control}}{Name of condition in the targets file which was the
  control [Default: control]} \item{\strong{annotation_file}}{Location of
  annotation file, see \link[deago]{annotateDataset} [Default: NULL]}
  \item{\strong{keep_images}}{Give a value of 1 to keep images which are used
  in the HTML report, see \link[deago]{plotToFile} for more information
  [Default: 0] } \item{\strong{qc_only}}{Give a value of 0 to run DESeq2
  analysis functions or a value of 1 to report only the QC plots [Default:
  1]} \item{\strong{go_analysis}}{Give a value of 1 to run GO term enrichment
  analysis using \link{topGO}. If value of 1 given for go_analysis, an
  annotation file must be provided as a value for the annotation key
  [Default: 0]} }

  Parameters are validated using \code{\link{validateConfig}}.

  \code{importConfig} returns a list containing the parameter values with the
  names corresponding to their keys.
}
\examples{
\dontrun{
 # Import tab-delimited config from working directory
 importConfig("config.txt")

 # Import tab-delimited config from another directory
 importConfig("config.txt", path="/path/to/config")

 # Import comma-delimited config from working directory
 importConfig("config.txt", sep=",")
}
}
\seealso{
Other configuration functions: \code{\link{buildConfig}},
  \code{\link{validateConfig}}, \code{\link{writeConfig}}

Other import functions: \code{\link{annotateDataset}},
  \code{\link{importAnnotation}},
  \code{\link{importTargets}}, \code{\link{readCountData}},
  \code{\link{validateConfig}},
  \code{\link{validateTargets}}
}
