% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importTargets.R
\name{importTargets}
\alias{importTargets}
\title{Read Targets File}
\usage{
importTargets(file, path = NULL, sep = "\\t")
}
\arguments{
\item{file}{A \link{character} string giving the name of the file.}

\item{path}{A \link{character} string giving the directory containing the
file. This may be omitted if the file is in the current working directory.}

\item{sep}{A \link{character} used to separate the fields.}
}
\value{
dataframe: containing the target information and labels
}
\description{
\code{importTargets} reads a delimited targets file into a
  dataframe.
}
\details{
The targets file is a text file which contains information about the
RNA samples, linking their experimental conditions to their count data files.
Rows correspond to samples and columns to their respective count data file
locations and the experimental conditions that were applied.  By default, it
is expected that the targets file will be tab-delimited but, other delimiters
can be used if specified using \code{sep}.

The targets file must contain at least three columns with the following,
column names:

\describe{ \item{\strong{filename}}{name of the file containing the count
data for the sample} \item{\strong{condition}}{experimental treatment or
condition that was applied} \item{\strong{replicate}}{replicate identifier -
can be numeric or character} }

\code{importTargets} adds a \code{label} column which contains labels
generated using the \code{condition} and \code{replicate} values for the
sample to use in plots.
}
\examples{
\dontrun{
 # Import tab-delimited targets file from working directory
 importTargets("targets.txt")

 # Import tab-delimited targets file from another directory
 importTargets("targets.txt", path="/path/to/targets")

 # Import comma-delimited targets file from working directory
 importTargets("targets.txt", sep=",")
}
}
\seealso{
Other target functions: \code{\link{validateTargets}}

Other import functions: \code{\link{annotateDataset}},
  \code{\link{importAnnotation}},
  \code{\link{importConfig}}, \code{\link{readCountData}},
  \code{\link{validateConfig}},
  \code{\link{validateTargets}}
}
\concept{import functions}
\concept{target functions}
