% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrastPlots.R
\name{labelDEgenes}
\alias{labelDEgenes}
\title{Label DE genes}
\usage{
labelDEgenes(contrast, geneSymbols = NULL, lfc = 2, alpha = 0.05)
}
\arguments{
\item{contrast}{A \link{DESeqResults} object}

\item{geneSymbols}{details}

\item{lfc}{Log2FoldChange threshold value}

\item{alpha}{FDR threshold}
}
\description{
Helper function to label DE genes from a DESeqResults dataframe
  with a given Log2FoldChange and padj cutoff. Adds a label column (de) where
  values are non-significant (0), up-regulated (1) or down-regulated (2).
}
