% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultsDirectory.R
\name{makeResultDir}
\alias{makeResultDir}
\title{Make Results Directory}
\usage{
makeResultDir(path = NULL, keep_images = 0)
}
\arguments{
\item{path}{A \code{\link{character}} string giving the location where the
results directory will be created.}

\item{keep_images}{A \code{\link{logical}} (default=0).}
}
\value{
name of the timestamped results directory
}
\description{
Creates a timestamped results directory.
}
\details{
\code{\link{makeResultDir}} creates a timestamped results
  directory in the current working directory or in the path specified using
  \code{path}.

  By default, when the HTML report is generated, the images used in the
  report are removed. If \code{keep_images} is set to 1 in the configuration
  file (see \code{\link{importConfig}}), then a folder called \code{images}
  is created as a subdirectory of the results directory.  All of the images
  used in the HTML report will then be stored in this subdirectory.
}
\examples{

 # With no image subdirectory
 makeResultDir(path=getwd())

 # With image subdirectory
 makeResultDir(path=getwd(), keep_images=1)

}
