% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcPlots.R
\name{pcaPlot}
\alias{pcaPlot}
\title{PCA Plot}
\usage{
pcaPlot(dds, resultsDir)
}
\arguments{
\item{dds}{DESeqDataSet}

\item{resultsDir}{character: path to timestamped results directory}
}
\description{
Generates a PCA plot. If images subdirectory exits, will also
  save the plot to images/pca.png.
}
