% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcPlots.R
\name{pcaScreePlot}
\alias{pcaScreePlot}
\title{PCA Scree Plot}
\usage{
pcaScreePlot(pc_list, resultsDir)
}
\arguments{
\item{pc_list}{List containing PC coordinates and PC summary from
getPrincipalComponents(dds)}

\item{resultsDir}{character: path to timestamped results directory}
}
\description{
Generates a PCA scree plot. If images subdirectory exits, will also
  save the plot to images/pca_scree.png.
}
