% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcPlots.R
\name{pcaSummary}
\alias{pcaSummary}
\title{PCA Summary}
\usage{
pcaSummary(pc_list)
}
\arguments{
\item{pc_list}{List containing PC coordinates and PC summary from
getPrincipalComponents(dds)}
}
\description{
Generates a summary table of PCA proportions and cumulative sum
  values
}
