% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrastPlots.R
\name{plotContrastMA}
\alias{plotContrastMA}
\title{Contrast MA plot}
\usage{
plotContrastMA(contrast, resultsDir, alpha = 0.05, lfc = 2,
  geneSymbols = NULL, geneLabels = FALSE)
}
\arguments{
\item{contrast}{A DESeqResults object}

\item{resultsDir}{Name of timestamped results directory}

\item{alpha}{FDR}

\item{lfc}{log2FoldChange}

\item{geneSymbols}{gene symbols}

\item{geneLabels}{description}
}
\description{
\code{\link{plotContrastMA}} generates a contrast MA plot.
}
\details{
Add some details here.
}
