% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcPlots.R
\name{plotCooks}
\alias{plotCooks}
\title{Plot Cook's Distances}
\usage{
plotCooks(dds, resultsDir)
}
\arguments{
\item{dds}{DESeqDataSet}

\item{resultsDir}{character: path to timestamped results directory}
}
\description{
Generates a boxplot showing Cook's distances per sample. If
  images subdirectory exits, will also save the plot to
  images/cooks_distances.png.
}
