% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcPlots.R
\name{plotDensity}
\alias{plotDensity}
\title{Plot density}
\usage{
plotDensity(dds, resultsDir)
}
\arguments{
\item{dds}{DESeqDataSet}

\item{resultsDir}{character: path to timestamped results directory}
}
\description{
Generates a desnity plot. If images subdirectory exits, will
  also save the plot to images/normalised_density.png.
}
