% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcPlots.R
\name{plotDispersionEstimates}
\alias{plotDispersionEstimates}
\title{Plot Dispersion Estimates}
\usage{
plotDispersionEstimates(dds, resultsDir)
}
\arguments{
\item{dds}{DESeqDataSet}

\item{resultsDir}{character: path to timestamped results directory}
}
\description{
Generates a scatter plot showing the dispersion estimates. If
  images subdirectory exits, will also save the plot to
  images/dispersion_estimates.png.
}
