% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcPlots.R
\name{plotSampleDistances}
\alias{plotSampleDistances}
\title{Plot Sample Distances}
\usage{
plotSampleDistances(dds, resultsDir)
}
\arguments{
\item{dds}{DESeqDataSet}

\item{resultsDir}{character: path to timestamped results directory}
}
\description{
Generates a heatmap of sample-to-sample distances with cluster
  dendrogram. If images subdirectory exits, will also save the plot to
  images/sample_distances.png.
}
