% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotToFile.R
\name{plotToFile}
\alias{plotToFile}
\title{Write Plot To File}
\usage{
plotToFile(plot, resultsDir, file, width = 600, height = 400)
}
\arguments{
\item{plot}{Plot to be written to file}

\item{resultsDir}{A \link{character} string giving the path to timestamped
results directory}

\item{file}{A \link{character} string giving name of file where the plot will
be saved}

\item{width}{An \link{integer} giving width of the device in pixels}

\item{height}{An \link{integer} giving height of the device in pixels}
}
\description{
\code{\link{plotToFile}} writes plot to PNG file.
}
\details{
If \code{keep_images} is set to 1 in the configuration file (see
  \code{\link{importConfig}}) then a subdirectory called \code{images} will
  be created in the timestamped results directory (see
  \code{\link{makeResultDir}}). \code{\link{plotToFile}} writes the
  designated plot to a PNG file within the images subdirectory.

  The dimensions of the plot, in pixels, can be defined using \code{width}
  and \code{height}.
}
