% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrastPlots.R
\name{plotVenn}
\alias{plotVenn}
\title{Venn diagram}
\usage{
plotVenn(countMatrix, resultsDir)
}
\arguments{
\item{countMatrix}{Count matrix preppared by vennCounts}

\item{resultsDir}{Name of timestamped results directory}
}
\description{
Plot venn diagram from venn count matrix
}
