% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseqAnalysis.R
\name{prepareColData}
\alias{prepareColData}
\title{Construct input matrix for DESeq2}
\usage{
prepareColData(targets, countdata, parameters)
}
\arguments{
\item{targets}{Name of object containing target data}

\item{countdata}{Name of object containing count data}

\item{parameters}{Parameters}
}
\value{
input matrix for DESeq2 analysis
}
\description{
Rows of coldata correspond to columns of countdata
}
