% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasts.R
\name{prepareContrast}
\alias{prepareContrast}
\title{Prepares contrasts table for writing}
\usage{
prepareContrast(dds, contrast)
}
\arguments{
\item{dds}{DESeqDataSet object containing DESeq2 analysis results}

\item{contrast}{DESeqResults object}
}
\description{
Takes results object and prepares dataframe for writing
}
