% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasts.R
\name{prepareContrastTable}
\alias{prepareContrastTable}
\title{Prepare contrast table}
\usage{
prepareContrastTable(contrast)
}
\arguments{
\item{contrast}{DESeqResults object}
}
\description{
Prepare a filtered table of contrast results containing DE genes only
}
