% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goAnalysis.R
\name{prepareGOtable}
\alias{prepareGOtable}
\title{Prepare GO results datatable}
\usage{
prepareGOtable(go_table)
}
\arguments{
\item{go_table}{A data frame containing topGO results}
}
\description{
Prepare GO results datatable
}
