% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{roundToBase}
\alias{roundToBase}
\title{Round To Base}
\usage{
roundToBase(x, base, direction)
}
\arguments{
\item{x}{An \link{integer} to be rounded}

\item{base}{An \link{integer} base value}

\item{direction}{A \link{character} value specifying the rounding directions
(can be either 'up' or 'down')}
}
\value{
An \link{integer} rounded to the nearest base value
}
\description{
\code{\link{roundToBase}} rounds value to nearest base.
}
\examples{
roundToBase(17, 2, 'up') # would give 18
roundToBase(17, 10, 'up') # would give 20
roundToBase(17, 10, 'down') # would give 10
roundToBase(-17, 10, 'down') # would give -20
}
