% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{roundToFactorOfTen}
\alias{roundToFactorOfTen}
\title{Round Factor Ten}
\usage{
roundToFactorOfTen(x)
}
\arguments{
\item{x}{An \link{integer} to be rounded}
}
\value{
An \link{integer} rounded to the nearest factor of ten
}
\description{
\code{\link{roundToFactorOfTen}} rounds value to nearest factor of 10.
}
\examples{
roundToFactorOfTen(19) # would give 20
roundToFactorOfTen(194) # would give 200
roundToFactorOfTen(1094) # would give 2000
}
