% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseqAnalysis.R
\name{runDESeqAnalysis}
\alias{runDESeqAnalysis}
\title{Differential expression analysis using DESeq2}
\usage{
runDESeqAnalysis(targets, countdata, parameters)
}
\arguments{
\item{targets}{Name of object containing target data}

\item{countdata}{Name of object containing count data}

\item{parameters}{Levels are re-ordered so that the level specified by ref is first}
}
\value{
Returns a DESeqDataSet object
}
\description{
Performs a default DESeq2 analysis
}
