% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goAnalysis.R
\name{runGOanalysis}
\alias{runGOanalysis}
\title{GO enrichment analysis}
\usage{
runGOanalysis(dds, contrasts, go_level_list)
}
\arguments{
\item{dds}{DESeq object}

\item{contrasts}{A \link{list}}

\item{go_level_list}{A \link{list}}
}
\description{
Run topGO enrichment anlaysis for contrast
}
