% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{validateConfig}
\alias{validateConfig}
\title{Validate Parameter List}
\usage{
validateConfig(parameters)
}
\arguments{
\item{parameters}{A \link{list} containing key/value pairs which define the
parameters for the analysis (see \link[deago]{importConfig} for more
information).}
}
\value{
A \link{list} containing the validated parameters.
}
\description{
\code{validateConfig} validates a list of parameter key/value
  pairs.
}
\details{
\code{validateConfig} takes a list of key/value pairs which
represent the parameters for the analysis, validates them and then returns
the validated parameter key/value list.

Required key/value pairs are: \describe{ \item{\strong{counts}}{Directory
where count data files are located.} \item{\strong{target}}{Location of
targets file which contains the mapping files and experimental conditions.}
\item{\strong{results_dir}}{Directory where results directory will be
created.} }

Other accepted key/value pairs and their defaults are: \describe{
\item{\strong{gene_ids}}{Name of column in count files which contains the
gene identifiers [Default: Geneid]} \item{\strong{alpha}}{[Significance
cut-off, see \link[DESeq2]{results} for more information Default: 0.05]}
\item{\strong{control}}{Name of condition in the targets file which was the
control [Default: control]} \item{\strong{annotation}}{Location of annotation
file, see \link[deago]{annotateDataset} [Default: NULL]}
\item{\strong{keep_images}}{Give a value of 1 to keep images which are used
in the HTML report, see \link[deago]{plotToFile} for more information
[Default: 0] } \item{\strong{qc_only}}{Give a value of 0 to run DESeq2
analysis functions or a value of 1 to report only the QC plots [Default: 1]}
\item{\strong{go_analysis}}{Give a value of 1 to run GO term enrichment
analysis using \link{topGO}. If value of 1 given for go_analysis, an
annotation file must be provided as a value for the annotation key [Default:
0]} }

\code{validateConfig} returns a list containing the validated parameter
values with the names corresponding to their keys.
}
\examples{
 # Basic parameter list
 params <- list ( 'counts' = "/path/to/counts.txt",
                  'target' = "/path/to/targets.txt",
                  'result_dir' = "/path/to/result_dir")

 # Validate parameters
 validateConfig(params)
}
\seealso{
Other configuration functions: \code{\link{buildConfig}},
  \code{\link{importConfig}}, \code{\link{writeConfig}}

Other import functions: \code{\link{annotateDataset}},
  \code{\link{importAnnotation}},
  \code{\link{importConfig}}, \code{\link{importTargets}},
  \code{\link{readCountData}},
  \code{\link{validateTargets}}
}
