% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importTargets.R
\name{validateTargets}
\alias{validateTargets}
\title{Validate Targets File}
\usage{
validateTargets(targets)
}
\arguments{
\item{targets}{A \link{data.frame} containing the target information.  See
\code{\link{importTargets}} for more information.}
}
\value{
If valid, returns \code{TRUE}.
}
\description{
Checks whether targets file contains the required columns.
}
\details{
\code{validateTargets} checks that the targets dataframe contains
  the three columns below:

  \describe{ \item{\strong{filename}}{name of the file containing the count
  data for the sample} \item{\strong{condition}}{experimental treatment or
  condition that was applied} \item{\strong{replicate}}{replicate identifier
  - can be numeric or character} }

  If the target dataframe contains the necessary columns,
  \code{validateTargets} returns \code{TRUE}.
}
\examples{

targets <- data.frame( 'filename'=c('file1','file2','file3','file4'),
                  'condition' = rep(c('A','B'), times=2),
                  'replicate' = c(1,2,1,2))

validateTargets(targets)
}
\seealso{
Other target functions: \code{\link{importTargets}}

Other import functions: \code{\link{annotateDataset}},
  \code{\link{importAnnotation}},
  \code{\link{importConfig}}, \code{\link{importTargets}},
  \code{\link{readCountData}}, \code{\link{validateConfig}}
}
