% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasts.R
\name{writeContrasts}
\alias{writeContrasts}
\title{Extract contrasts from DESeq2 results}
\usage{
writeContrasts(dds, contrasts, resultsDir)
}
\arguments{
\item{dds}{A \link[DESeq2]{DESeqDataSet} object.}

\item{contrasts}{A \link{list}}

\item{resultsDir}{A \link{character} string giving the path to timestamped
results directory}
}
\description{
Extract contrasts from DESeq2 results and write to file
}
\details{
The details
}
