% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goAnalysis.R
\name{writeGOtables}
\alias{writeGOtables}
\title{Write GO results table to file}
\usage{
writeGOtables(go_tables, resultsDir)
}
\arguments{
\item{go_tables}{A list containing topGO results tables}

\item{resultsDir}{A \link{character} string giving the path to timestamped
results directory}
}
\description{
Write GO results table to file
}
\details{
The details
}
