% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goAnalysis.R
\name{getGOidentifiers}
\alias{getGOidentifiers}
\title{Get gene identifiers for GO terms}
\usage{
getGOidentifiers(contrast, GOdata, GOtable, gene_level = "all")
}
\arguments{
\item{contrast}{DESeqResults object}

\item{GOdata}{topGO object}

\item{GOtable}{topGO table}

\item{gene_level}{all, up or down [all]}
}
\description{
Get gene identifiers for GO terms
}
