% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goAnalysis.R
\name{getGOsymbols}
\alias{getGOsymbols}
\title{Get gene symbols for GO terms}
\usage{
getGOsymbols(contrast, GOdata, GOtable, gene_level = "all")
}
\arguments{
\item{contrast}{DESeqResults object}

\item{GOdata}{topGO object}

\item{GOtable}{topGO table}

\item{gene_level}{all, up or down [all]}
}
\description{
Get gene symbols for GO terms
}
