% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goAnalysis.R
\name{prepareGOdata}
\alias{prepareGOdata}
\title{Prepare GO data}
\usage{
prepareGOdata(dds, contrast, go_level, gene_level = "all")
}
\arguments{
\item{dds}{DESeq object}

\item{contrast}{DESeqResults object}

\item{go_level}{BP or MF}

\item{gene_level}{all, up or down [all]}
}
\description{
Prepare GO terms for TopGO analysis
}
