% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goAnalysis.R
\name{prepareGOdata}
\alias{prepareGOdata}
\title{Prepare GO data}
\usage{
prepareGOdata(dds, contrast, go_level)
}
\arguments{
\item{dds}{DESeq object}

\item{contrast}{DESeqResults object}

\item{go_level}{BP or MF}
}
\description{
Prepare GO terms for TopGO analysis
}
