% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goAnalysis.R
\name{prepareGOtable}
\alias{prepareGOtable}
\title{GO enrichment analysis}
\usage{
prepareGOtable(dds, contrast, go_level)
}
\arguments{
\item{dds}{DESeq object}

\item{contrast}{DESeqResults object}

\item{go_level}{BP or MF}
}
\description{
Run topGO enrichment anlaysis for contrast
}
